+++ js/main.js
@@ -1,89 +1,3 @@
-const COUNT_PHOTOS = 25;
-const MIN_LIKES = 15;
-const MAX_LIKES = 200;
-const MIN_AVATAR_COUNT = 1;
-const MAX_AVATAR_COUNT = 6;
-const COUNT_COMMENTS = 5;
-const COUNT_MESSAGES = 2;
+import {generatePhotos} from './data.js';
 
-const NAMES = [
-  'Иван',
-  'Хуан Себастьян',
-  'Мария',
-  'Кристоф',
-  'Виктор',
-  'Юлия',
-  'Анна',
-  'Люси',
-  'Люпита',
-  'Вашингтон',
-];
-
-const DESCRIPTIONS = [
-  'описание фотографии 1',
-  'описание фотографии 2',
-  'описание фотографии 3',
-  'описание фотографии 4',
-  'описание фотографии 5'
-];
-
-const MESSAGES = [
-  'Всё отлично!',
-  'В целом всё неплохо. Но не всё.',
-  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
-  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
-  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
-  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
-];
-
-const getRandomInteger = (min, max) => {
-  const lower = Math.ceil(Math.min(min, max));
-  const upper = Math.floor(Math.max(min, max));
-  const result = Math.random() * (upper - lower + 1) + lower;
-  return Math.floor(result);
-};
-
-const generateMessages = (number) => {
-  let messages = '';
-  for (let i = 1; i <= number; i++) {
-    messages += MESSAGES[getRandomInteger(0, MESSAGES.length - 1)];
-  }
-
-  return messages;
-};
-
-const createComment = (number, indexPhoto) => ({
-  // Можно ли выполнить так, генерацию уникальных идентификаторов комментария ? Судя по ТЗ это не должно противоречить ему
-  id: (indexPhoto * 10 + number),
-  avatar: `img/avatar-${getRandomInteger(MIN_AVATAR_COUNT, MAX_AVATAR_COUNT)}.svg`,
-  message: generateMessages(getRandomInteger(1, COUNT_MESSAGES)),
-  name: NAMES[getRandomInteger(0, NAMES.length - 1)],
-});
-
-const generateComments = (number) => {
-  const comments = [];
-  for (let i = 1; i <= getRandomInteger(1, COUNT_COMMENTS); i++) {
-    comments.push(createComment(i, number));
-  }
-
-  return comments;
-};
-
-const createPhoto = (number) => ({
-  id: number,
-  url: `photos/${number}.jpg`,
-  description: DESCRIPTIONS[getRandomInteger(0, DESCRIPTIONS.length - 1)],
-  likes: getRandomInteger(MIN_LIKES, MAX_LIKES),
-  comments: generateComments(number),
-});
-
-const generatePhotos = () => {
-  const photos = [];
-  for (let i = 1; i <= COUNT_PHOTOS; i++) {
-    photos.push(createPhoto(i));
-  }
-
-  return photos;
-};
-
-generatePhotos();
+const allPhotos = generatePhotos();
